fix: casts cached metrics to float
Redis returns numerics unserialized as strings on cache hits, violating the float return type. Casts the remembered value to restore the contract across cache stores.

Closes #14.

+++ src/Calculators/CachingCalculator.php
@@ -49,6 +49,6 @@ final readonly class CachingCalculator implements StorageCalculator
      */
     private function remember(string $metric, Closure $callback): float
     {
-        return Cache::remember("{$this->cacheKey}:{$metric}", $this->ttl, $callback);
+        return (float) Cache::remember("{$this->cacheKey}:{$metric}", $this->ttl, $callback);
     }
 }
